# fix(@mastra/core, @mastra/memory, @mastra/pg): semantic recall perf + memory bundle size