--- a/.gitignore
+++ b/.gitignore
@@ -19,6 +19,7 @@ openapi-ts-error*
 
 .mastra
 memory.db
+memory.db-*
 
 # github actions
 .secrets