--- a/packages/core/src/memory/memory.ts
+++ b/packages/core/src/memory/memory.ts
@@ -140,22 +140,15 @@ export abstract class MastraMemory extends MastraBase {
     return {};
   }
 
-  protected async createEmbeddingIndex(providedDimension?: number): Promise<{ indexName: string }> {
+  protected async createEmbeddingIndex(dimensions?: number): Promise<{ indexName: string }> {
     const defaultDimensions = 1536;
-    // AI SDK doesn't expose a way to check how many dimensions a model uses.
-    const dimensionsByModelId: Record<string, number> = {
-      'bge-small-en-v1.5': 384,
-      'bge-base-en-v1.5': 768,
-      'voyage-3-lite': 512,
-    };
-
-    const dimensions = dimensionsByModelId[this.embedder.modelId] || defaultDimensions;
     const isDefault = dimensions === defaultDimensions;
-    const indexName = isDefault ? 'memory_messages' : `memory_messages_${dimensions}`;
+    const usedDimensions = dimensions ?? defaultDimensions;
+    const indexName = isDefault ? 'memory_messages' : `memory_messages_${usedDimensions}`;
 
     await this.vector.createIndex({
-      indexName: typeof providedDimension !== `undefined` ? `memory_messages_${providedDimension}` : indexName,
-      dimension: dimensions,
+      indexName,
+      dimension: usedDimensions,
     });
     return { indexName };
   }