… .changeset/bumpy-words-lead.md
… packages/core/package.json
… packages/memory/package.json
… pnpm-lock.yaml
… stores/pg/package.json
… stores/pg/src/vector/index.ts